second commit

+++ user/delete.php
@@ -1,6 +1,13 @@
-<?php
-include('../include/db.php');
-$id = $_GET['id'];
-mysqli_query($conn, "DELETE FROM user WHERE id = $id") or die('Query Error');
+<script>
+  var result = confirm("Press a button!");
+  if (result == true) {
+    <?php
+    include('../include/db.php');
 
-header('Location: ../dashboard.php');
+    $id = $_GET['id'];
+    mysqli_query($conn, "DELETE FROM user WHERE id = $id") or die('Query Error');
+
+    header('Location: index.php');
+    ?>
+  }
+</script>
